[FLINK-24500][table] Move SqlDateTimeUtils to table-common as DateTimeUtils

This closes #17454.
